time/energy in message data only integer precision

+++ brouter-core/src/main/java/btools/router/MessageData.java
@@ -56,8 +56,8 @@ final class MessageData implements Cloneable
          + "\t" + linkinitcost
          + "\t" + wayKeyValues
          + "\t" + ( nodeKeyValues == null ? "" : nodeKeyValues )
-            + "\t" + time
-            + "\t" + energy;
+            + "\t" + ((int)time)
+            + "\t" + ((int)energy);
   }
 
   void add( MessageData d )
